refactor: [12] init 파일 경로 수정
